[Refactor] Cleanup old/unused files
