--- app/src/main/java/android/content/res/ResourceId.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (C) 2020 Muntashir Al-Islam
- * Copyright (C) 2017 The Android Open Source Project
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package android.content.res;
-
-import androidx.annotation.AnyRes;
-
-/**
- * Provides a set of utility methods for dealing with Resource IDs.
- */
-public final class ResourceId {
-    /**
-     * Checks whether the integer {@code id} is a valid resource ID, as generated by AAPT.
-     * <p>Note that a negative integer is not necessarily an invalid resource ID, and custom
-     * validations that compare the {@code id} against {@code 0} are incorrect.</p>
-     * @param id The integer to validate.
-     * @return {@code true} if the integer is a valid resource ID.
-     */
-    public static boolean isValid(@AnyRes int id) {
-        // With the introduction of packages with IDs > 0x7f, resource IDs can be negative when
-        // represented as a signed Java int. Some legacy code assumes -1 is an invalid resource ID,
-        // despite the existing documentation.
-        return id != -1 && (id & 0xff000000) != 0 && (id & 0x00ff0000) != 0;
-    }
-}
