+++ app/src/main/java/io/github/muntashirakon/AppManager/misc/DeviceInfo2.java
@@ -14,6 +14,7 @@ import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.UserHandleHidden;
 import android.text.SpannableStringBuilder;
+import android.text.TextUtils;
 import android.text.format.DateFormat;
 import android.text.format.Formatter;
 import android.util.DisplayMetrics;
@@ -28,8 +29,6 @@ import androidx.collection.ArrayMap;
 import androidx.core.os.LocaleListCompat;
 import androidx.core.util.Pair;
 import androidx.fragment.app.FragmentActivity;
-
-import com.android.internal.util.TextUtils;
 
 import java.security.Provider;
 import java.security.Security;
@@ -47,6 +46,7 @@ import io.github.muntashirakon.AppManager.runner.Runner;
 import io.github.muntashirakon.AppManager.runner.RunnerUtils;
 import io.github.muntashirakon.AppManager.users.UserInfo;
 import io.github.muntashirakon.AppManager.users.Users;
+import io.github.muntashirakon.AppManager.utils.TextUtilsCompat;
 import io.github.muntashirakon.AppManager.utils.Utils;
 import io.github.muntashirakon.proc.ProcFs;
 import io.github.muntashirakon.util.LocalizedString;
@@ -180,7 +180,7 @@ public class DeviceInfo2 implements LocalizedString {
             securityProviders.add(provider.getName() + " (v" + provider.getVersion() + ")");
         }
         builder.append(getStyledKeyValue(ctx, R.string.security_providers,
-                TextUtils.joinSpannable(", ", securityProviders))).append("\n");
+                TextUtilsCompat.joinSpannable(", ", securityProviders))).append("\n");
         // CPU info
         builder.append("\n").append(getTitleText(ctx, R.string.cpu)).append("\n");
         if (cpuHardware != null) {
@@ -219,7 +219,7 @@ public class DeviceInfo2 implements LocalizedString {
             localeStrings.add(systemLocales.get(i).getDisplayName());
         }
         builder.append("\n").append(getTitleText(ctx, R.string.languages))
-                .append("\n").append(TextUtils.joinSpannable(", ", localeStrings))
+                .append("\n").append(TextUtilsCompat.joinSpannable(", ", localeStrings))
                 .append("\n");
         if (users != null) {
             // Users
@@ -230,7 +230,7 @@ public class DeviceInfo2 implements LocalizedString {
                 userNames.add(user.name);
             }
             builder.append(String.format(Locale.getDefault(), "%d", users.size())).append(" (")
-                    .append(TextUtils.joinSpannable(", ", userNames))
+                    .append(TextUtilsCompat.joinSpannable(", ", userNames))
                     .append(")\n");
             // App stats per user
             builder.append("\n").append(getTitleText(ctx, R.string.apps)).append("\n");
@@ -271,7 +271,7 @@ public class DeviceInfo2 implements LocalizedString {
                 } else featureStrings.add(info.name);
             }
         }
-        builder.append(TextUtils.joinSpannable("\n", featureStrings)).append("\n");
+        builder.append(TextUtilsCompat.joinSpannable("\n", featureStrings)).append("\n");
         return builder;
     }
