+++ app/src/main/java/io/github/muntashirakon/AppManager/sysconfig/SystemConfig.java
@@ -8,6 +8,7 @@ import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Environment;
 import android.os.Process;
+import android.text.TextUtils;
 import android.util.ArrayMap;
 import android.util.SparseArray;
 import android.util.Xml;
@@ -16,8 +17,6 @@ import androidx.annotation.IntRange;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.WorkerThread;
-
-import com.android.internal.util.TextUtils;
 
 import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserException;
@@ -35,6 +34,7 @@ import io.github.muntashirakon.AppManager.logs.Log;
 import io.github.muntashirakon.AppManager.misc.OsEnvironment;
 import io.github.muntashirakon.AppManager.misc.SystemProperties;
 import io.github.muntashirakon.AppManager.utils.ArrayUtils;
+import io.github.muntashirakon.AppManager.utils.TextUtilsCompat;
 import io.github.muntashirakon.compat.xml.XmlUtils;
 import io.github.muntashirakon.io.IoUtils;
 import io.github.muntashirakon.io.Path;
@@ -1170,10 +1170,10 @@ public class SystemConfig {
                     }
                     break;
                     case SysConfigType.TYPE_NAMED_ACTOR: {
-                        String namespace = TextUtils.safeIntern(
+                        String namespace = TextUtilsCompat.safeIntern(
                                 parser.getAttributeValue(null, "namespace"));
                         String actorName = parser.getAttributeValue(null, "name");
-                        String pkgName = TextUtils.safeIntern(
+                        String pkgName = TextUtilsCompat.safeIntern(
                                 parser.getAttributeValue(null, "package"));
                         if (TextUtils.isEmpty(namespace)) {
                             Log.e(TAG, "<" + name + "> without namespace in " + permFile
